requirements.txt
flask
transformers
pillow
torch
torchvision